Test loading extended ontology
