+++ indra/tests/test_ontology.py
@@ -1,7 +1,5 @@
-import copy
 from indra.statements import Agent
 from indra.ontology.bio import bio_ontology
-from indra.ontology.world import world_ontology
 from indra.databases import go_client, hgnc_client
 from indra.ontology.standardize import \
     standardize_agent_name, standardize_db_refs, standardize_name_db_refs
@@ -133,19 +131,6 @@ def test_chebi_isa():
 #    c1 = ent_hierarchy.components[uri_prkag1]
 #    c2 = ent_hierarchy.components[uri_ampk]
 #    assert c1 == c2
-
-
-def test_hm_opposite_polarity():
-    concept1 = 'wm/concept/causal_factor/food_insecurity/food_instability'
-    concept2 = 'wm/concept/causal_factor/food_security/food_stability'
-    concept3 = ('wm/concept/causal_factor/environmental/meteorologic/'
-                'precipitation/flooding')
-    assert world_ontology.is_opposite('WM', concept1, 'WM', concept2)
-    assert world_ontology.is_opposite('WM', concept2, 'WM', concept1)
-    assert not world_ontology.is_opposite('WM', concept1, 'WM', concept3)
-    assert world_ontology.get_polarity('WM', concept1) == -1
-    assert world_ontology.get_polarity('WM', concept2) == 1
-    assert world_ontology.get_polarity('UN', 'something') is None
 
 
 def test_name_standardize_hgnc_up():
@@ -315,18 +300,6 @@ def test_standardize_chembl():
     db_refs = standardize_db_refs({'DRUGBANK': 'DB00305'})
     assert 'CHEMBL' in db_refs, db_refs
     assert db_refs['CHEMBL'] == 'CHEMBL105', db_refs
-
-
-def test_world_ontology_add_entry():
-    ont = copy.deepcopy(world_ontology)
-    nat_dis = ('wm/concept/causal_factor/crisis_and_disaster/'
-               'environmental_disasters/natural_disaster')
-
-    new_node = nat_dis + '/floods'
-    assert not ont.isa('WM', new_node, 'WM', nat_dis)
-    ont.add_entry(new_node, examples=['floods'])
-    assert ont.isa('WM', new_node, 'WM', nat_dis)
-    ont_yml = ont.dump_yml_str()
 
 
 def test_efo_bfo_relations():
